--- community/plugins/spring-ai-alibaba-starter-gaode-map/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,0 @@
-com.alibaba.cloud.ai.GaoDeConfig
