+++ community/plugins/spring-ai-alibaba-starter-plugin-gaode/src/main/java/com/alibaba/cloud/ai/plugin/gaode/GaoDeConfiguration.java
@@ -1,4 +1,4 @@
-package com.alibaba.cloud.ai.plugin.dingtalk;
+package com.alibaba.cloud.ai.plugin.gaode;
 /*
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
@@ -15,6 +15,7 @@ package com.alibaba.cloud.ai.plugin.dingtalk;
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import com.alibaba.cloud.ai.plugin.gaode.function.WeatherSearchFunction;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
@@ -26,14 +27,14 @@ import org.springframework.context.annotation.Description;
  * @author YunLong
  */
 @Configuration
-@EnableConfigurationProperties(DingTalkProperties.class)
-@ConditionalOnProperty(prefix = "spring.ai.alibaba.plugin.dingtalk", name = "enabled", havingValue = "true")
-public class DingTalkConfiguration {
+@EnableConfigurationProperties(GaoDeProperties.class)
+@ConditionalOnProperty(prefix = "spring.ai.alibaba.plugin.gaode", name = "enabled", havingValue = "true")
+public class GaoDeConfiguration {
 
     @Bean
     @ConditionalOnMissingBean
-    @Description("Send DingTalk group chat messages using a custom robot")
-    public DingTalkService dingTalkGroupSendMessageByCustomRobotFunction(DingTalkProperties dingTalkProperties) {
-        return new DingTalkService(dingTalkProperties);
+    @Description("Get weather information according to address.")
+    public WeatherSearchFunction gaoDeGetAddressWeatherFunction(GaoDeProperties gaoDeProperties) {
+        return new WeatherSearchFunction(gaoDeProperties);
     }
 }
