+++ community/plugins/spring-ai-alibaba-starter-plugin-gaode/src/main/java/com/alibaba/cloud/ai/plugin/gaode/function/WeatherSearchFunction.java
@@ -1,7 +1,22 @@
-package com.alibaba.cloud.ai.function;
-
-import com.alibaba.cloud.ai.properties.GaoDeProperties;
-import com.alibaba.cloud.ai.service.WebService;
+package com.alibaba.cloud.ai.plugin.gaode.function;
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import com.alibaba.cloud.ai.plugin.gaode.GaoDeProperties;
+import com.alibaba.cloud.ai.plugin.gaode.service.WebService;
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.JSONArray;
 import com.alibaba.fastjson.JSONObject;
