+++ community/plugins/spring-ai-alibaba-starter-plugin-gaode/src/main/java/com/alibaba/cloud/ai/plugin/gaode/service/WebService.java
@@ -1,8 +1,22 @@
-package com.alibaba.cloud.ai.service;
+package com.alibaba.cloud.ai.plugin.gaode.service;
 
-import com.alibaba.cloud.ai.properties.GaoDeProperties;
-import org.springframework.stereotype.Service;
-
+import com.alibaba.cloud.ai.plugin.gaode.GaoDeProperties;
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 import java.net.URI;
 import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
@@ -13,7 +27,7 @@ import java.util.concurrent.CompletableFuture;
 /**
  * @author YunLong
  */
-@Service
+
 public class WebService {
 
     private final String baseUrl = "https://restapi.amap.com/v3";
