+++ pom.xml
@@ -38,6 +38,7 @@
 		<module>community/plugins/spring-ai-alibaba-starter-plugin-baidusearch</module>
 		<module>community/plugins/spring-ai-alibaba-starter-plugin-bingsearch</module>
 		<module>community/plugins/spring-ai-alibaba-starter-plugin-dingtalk</module>
+		<module>community/plugins/spring-ai-alibaba-starter-plugin-gaode</module>
     </modules>
 
 	<properties>
