+++ spring-ai-alibaba-examples/function-calling-example/pom.xml
@@ -78,6 +78,12 @@
 			<version>${spring-ai-alibaba.version}</version>
 		</dependency>
 
+		<dependency>
+			<groupId>com.alibaba.cloud.ai</groupId>
+			<artifactId>spring-ai-alibaba-starter-plugin-gaode</artifactId>
+			<version>${spring-ai-alibaba.version}</version>
+		</dependency>
+
 	</dependencies>
 
 	<build>
